refactor(core): extract frame/page schema into longfist/core.h leaf
Move frame_header (tag 0), page_header (tag 1) and the two enums the journal
spine needs (FrameDataType, PageStatus), plus the KF_JSON_SERIALIZE_ENUM helper,
out of types.h / enums.h into a new schema leaf header longfist/core.h. Business
types (Order/Trade/...) stay in types.h and are not visible from the leaf, so a
pure journal core can include only the schema it needs.

Type tags and field layout are moved verbatim -- on-disk mmap journals stay
byte-compatible. Pure refactor: types.h/enums.h include the leaf, AllTypes and
all behaviour are unchanged. libkungfu builds green.
