+++ framework/core/src/include/kungfu/longfist/types.h
@@ -54,48 +54,8 @@ KF_DEFINE_MARK_TYPE(KeepPositionsRequest, 10453);
 KF_DEFINE_MARK_TYPE(RebuildPositionsRequest, 10454);
 KF_DEFINE_MARK_TYPE(SocketData, 10751);
 
-KF_DEFINE_PACK_TYPE(                                           //
-    frame_header, 0, PK(gen_time), TIMESTAMP(gen_time),        //
-    /** total frame length (including header and data body);                //
-     *  ADR-0001: serves as the frame publication token. Written last with   //
-     *  std::atomic_ref release by the writer and read with acquire by the    //
-     *  reader (see frame.h). NOT volatile: volatile gives no cross-thread    //
-     *  ordering on weak-memory (ARM) targets. */                            //
-    (uint32_t, length),                                        //
-    /** header length */                                       //
-    (uint32_t, header_length),                                 //
-    /** generate time of the frame data */                     //
-    (int64_t, gen_time),                                       //
-    /** trigger time for this frame, use for latency stats */  //
-    (int64_t, trigger_time),                                   //
-    /** msg type of the data in frame (ADR-0001: no longer volatile;          //
-     *  visibility is guaranteed by the length release/acquire token) */      //
-    (int32_t, msg_type),                                       //
-    /** source of this frame */                                //
-    (uint32_t, source),                                        //
-    /** dest of this frame */                                  //
-    (uint32_t, dest),                                          //
-    /** json or raw struct */                                  //
-    (enums::FrameDataType, data_type),                         //
-    /** the real writer of this frame */                       //
-    (uint32_t, initial_source),                                //
-    /** key of frame */                                        //
-    (uint64_t, frame_uid),                                     //
-    /** current_frame of reader when generate this frame */    //
-    (uint64_t, trigger_frame_uid),                             //
-    /** stream_id */                                           //
-    (uint64_t, stream_id)                                      //
-);
-
-KF_DEFINE_PACK_TYPE(                          //
-    page_header, 1, PK(version), PERPETUAL(), //
-    (uint32_t, version),                      //
-    (uint32_t, page_header_length),           //
-    (uint64_t, page_size),                    //
-    (uint32_t, frame_header_length),          //
-    (longfist::enums::PageStatus, status),    // 0 close 1 preopen 2 open 3 flushing
-    (uint64_t, last_frame_position)           //
-);
+// frame_header(tag 0) 与 page_header(tag 1) 已移入 schema 叶子 kungfu/longfist/core.h
+// （经 enums.h -> core.h 在此可见，type tag 与字段布局不变）。
 
 KF_DEFINE_PACK_TYPE(                         //
     Asset, 101, PK(holder_uid), PERPETUAL(), //
